requirements.txt
modulegraph~=0.19.2
py2app~=0.28.2
macholib~=1.16
setuptools~=57.0.0
altgraph~=0.17.2
PySimpleGUI~=4.60.1
pip~=21.1.2
wheel~=0.36.2
Pillow~=9.1.1
click~=8.1.3
gTTS~=2.2.4
six~=1.16.0
urllib3~=1.26.9
requests~=2.28.0
certifi~=2022.6.15
idna~=3.3
playsound~=1.3.0
opencv-python~=4.6.0.66
numpy~=1.22.4
astropy~=5.1
PyYAML~=6.0
packaging~=21.3
pyparsing~=3.0.9
cffi~=1.15.0
cryptography~=37.0.2
paramiko~=2.11.0
bcrypt~=3.2.2
pycparser~=2.21
mysql-connector-python~=8.0.29
protobuf~=4.21.1